src/post-service/requirements.txt
grpcio>=1.64.0
grpcio-tools>=1.64.0
psycopg2-binary>=2.9.9
